requirements.txt
torch>=2.1.0
transformers>=4.37.0
sentence-transformers>=2.3.0
faiss-cpu>=1.7.4
networkx>=3.2.1
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.1.0
scikit-learn>=1.3.0
datasets>=2.16.0
pyyaml>=6.0.1
tqdm>=4.66.0
matplotlib>=3.8.0
seaborn>=0.13.0
pyvis>=0.3.2
nltk>=3.8.1
spacy>=3.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
wandb>=0.16.0
pytest>=7.4.0
hydra-core>=1.3.0
omegaconf>=2.3.0
pgmpy>=0.1.25